requirements.txt
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
numpy==1.18.2
PyYAML==5.3.1
sentencepiece==0.1.8
# tensorboard==1.12.2
tensorflow-gpu==1.12.0
